Revert "move network commands to network namespace"
This reverts commit 1db3b3477cbae191c23466fca6820a2bf69869b1.
