+++ lib/api/client-commands/captureNetworkRequests.js
@@ -1,5 +1,5 @@
-const ClientCommand = require('../_base-command.js');
-const {Logger} = require('../../../utils/index.js');
+const ClientCommand = require('./_base-command.js');
+const {Logger} = require('../../utils');
 
 /**
  * Capture outgoing network calls from the browser.
@@ -27,10 +27,6 @@ const {Logger} = require('../../../utils/index.js');
  * @moreinfo nightwatchjs.org/guide/network-requests/capture-network-calls.html
  */
 class CaptureNetworkCalls extends ClientCommand {
-
-  static get namespacedAliases() {
-    return 'captureNetworkRequests';
-  }
 
   performAction(callback) {
