+++ lib/api/client-commands/setNetworkConditions.js
@@ -1,5 +1,5 @@
-const ClientCommand = require('../_base-command.js');
-const {Logger} = require('../../../utils/index.js');
+const ClientCommand = require('./_base-command.js');
+const {Logger} = require('../../utils');
 
 /**
  * 
@@ -23,11 +23,6 @@ const {Logger} = require('../../../utils/index.js');
  * @api protocol.sessions
  */
 class SetNetworkConditions extends ClientCommand {
-
-  static get namespacedAliases() {
-    return 'setNetworkConditions';
-  }
-
   performAction(callback) {
     if (!this.api.isChrome() && !this.api.isEdge()) {
       const error = new Error('SetNetworkConditions is not supported while using this driver');
